moved/cleaned up utils, fixed write_file
